+++ apps/web/src/app/live/[organizationHandle]/layout.tsx
@@ -4,6 +4,7 @@ import { CustomerViewStateProvider, SpinProvider } from "./controller"
 import { WheelDisplay } from "./WheelDisplay"
 import { InformationDisplay } from './InformationDisplay'
 import { resolveOrganization } from './serverActions'
+import { cn } from '@repo/utilities'
 
 
 
@@ -42,13 +43,18 @@ export default async function Layout({
         <SpinProvider>
           {/* TODO: Find a good background video for the wheel */}
           {/* <video autoPlay muted loop playsInline
-            className="bg-slate-200 absolute -z-10 top-1/2 left-1/2 w-full h-full object-cover -translate-x-1/2 -translate-y-1/2"
+            className="absolute  top-1/2 left-1/2 w-full h-full object-cover -translate-x-1/2 -translate-y-1/2"
           > <source src="/videos/tunnel.mp4" type="video/mp4" />
           </video> */}
-          <WheelDisplay />
-          <InformationDisplay>
-            {children}
-          </InformationDisplay>
+          <div className={cn(
+            "grid grid-rows-[1fr] md:grid-cols-[1fr_minmax(542px,1fr)] h-[100svh]",
+            "w-full overflow-clip relative"
+          )}>
+            <WheelDisplay />
+            <InformationDisplay>
+              {children}
+            </InformationDisplay>
+          </div>
         </SpinProvider>
       </CustomerViewStateProvider>
     </div>